video_transcription.py
"""
Video transcription module for converting audio/video to text and processing with AI
"""

def process(input_file, raw_text_output, processed_text_output):
    """
    Transcribe audio/video file and process the text
    
    Args:
        input_file (str): Path to the input audio or video file
        raw_text_output (str): Path where the raw transcription should be saved
        processed_text_output (str): Path where the processed text should be saved
    
    Returns:
        bool: True if processing was successful, False otherwise
    """
    try:
        # This is a placeholder implementation
        # In a real implementation, you would:
        # 1. Use Whisper or similar to transcribe the audio/video
        # 2. Process the transcription with an AI model to organize the content
        # 3. Save both raw and processed versions to Word documents
        
        # For now, we'll just create simple Word documents with placeholder data
        from docx import Document
        
        # Create raw transcription document
        raw_doc = Document()
        raw_doc.add_heading('Raw Transcription', 0)
        raw_doc.add_paragraph(f'Input file: {input_file}')
        raw_doc.add_paragraph('This is a placeholder for the raw transcription.')
        raw_doc.add_paragraph('In a real implementation, this would contain the actual transcribed text.')
        raw_doc.save(raw_text_output)
        
        # Create processed document
        processed_doc = Document()
        processed_doc.add_heading('Processed Meeting Notes', 0)
        processed_doc.add_paragraph(f'Input file: {input_file}')
        processed_doc.add_paragraph('This is a placeholder for the AI-processed meeting notes.')
        processed_doc.add_paragraph('In a real implementation, this would contain organized meeting notes.')
        processed_doc.add_heading('Key Points', 1)
        processed_doc.add_paragraph('• Placeholder for key point 1')
        processed_doc.add_paragraph('• Placeholder for key point 2')
        processed_doc.add_paragraph('• Placeholder for key point 3')
        processed_doc.add_heading('Action Items', 1)
        processed_doc.add_paragraph('• Placeholder for action item 1')
        processed_doc.add_paragraph('• Placeholder for action item 2')
        processed_doc.save(processed_text_output)
        
        return True
    except Exception as e:
        print(f"Error during transcription processing: {e}")
        return False