requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.1
mysql-connector-python>=8.0.26
openpyxl>=3.0.7
python-docx>=0.8.11